Move CacheTest to kinto.core.cache.testing
